Import version 88.03
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@49 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/button.h
@@ -72,7 +72,7 @@ private:
 
 
 public:
-    int background; //@author hsiegeln
+    uint8 background; //@author hsiegeln
 
 
     /**
@@ -92,7 +92,7 @@ public:
     bool pressed;
 
 
-    int kennfarbe;
+    uint8 kennfarbe;
 
     button_t(const button_t & other);
